Extract logic to redirect to login page
