+++ frontend/pages/institutions/create.tsx
@@ -1,60 +1,31 @@
-import { useQuery } from "@apollo/client/react";
-import { stringifyApolloError } from "../../lib/apollo";
-import { NextRouter, useRouter } from "next/router";
 import { Skeleton, Row, Col, Card, App } from "antd";
 import Layout from "../../components/Layout";
 import paths from "../../paths";
-import { useEffect } from "react";
-import { CurrentUserDocument } from "../../queries/currentUser.generated";
 import CreateInstitution from "../../components/institutions/CreateInstitution";
-
-function redirectToLoginPage(router: NextRouter): void {
-  router.push({
-    pathname: paths.openIdConnectClientLogin,
-    query: { returnTo: paths.institutionCreate },
-  });
-}
+import { useRequireAuth } from "../../lib/hooks/useRequireAuth";
 
 function Page() {
-  const router = useRouter();
+	const { currentUser } = useRequireAuth({ returnTo: paths.institutionCreate });
 
-  const { loading, data, error } = useQuery(CurrentUserDocument);
-  const currentUser = data?.currentUser;
-  const shouldRedirect = !(loading || error || currentUser);
-  const { message } = App.useApp();
+	if (!currentUser) {
+		return (
+			<Layout>
+				<Skeleton active avatar title />
+			</Layout>
+		);
+	}
 
-  useEffect(() => {
-    if (error) {
-      message.error(stringifyApolloError(error));
-    }
-  }, [error]);
-
-  useEffect(() => {
-    if (router.isReady && shouldRedirect) {
-      redirectToLoginPage(router);
-    }
-  }, [shouldRedirect, router]);
-
-  if (loading || !currentUser) {
-    // TODO Handle this case properly.
-    return (
-      <Layout>
-        <Skeleton active avatar title />
-      </Layout>
-    );
-  }
-
-  return (
-    <Layout>
-      <Row justify="center">
-        <Col>
-          <Card title="Create">
-            <CreateInstitution ownerIds={[currentUser.uuid]} />
-          </Card>
-        </Col>
-      </Row>
-    </Layout>
-  );
+	return (
+		<Layout>
+			<Row justify="center">
+				<Col>
+					<Card title="Create">
+						<CreateInstitution ownerIds={[currentUser.uuid]} />
+					</Card>
+				</Col>
+			</Row>
+		</Layout>
+	);
 }
 
 export default Page;
